pyproject.toml
[project]
name = "browser-use-script"
description = "Enhanced CLI tool for browser automation using browser-use library"
authors = [{ name = "PierrunoYT" }]
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "browser-use>=0.2.2",
    "langchain-openai>=0.3.11",
    "langchain-anthropic>=0.3.3",
    "langchain-core>=0.3.49",
    "langchain>=0.3.21",
    "playwright>=1.52.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.4,<2.11.0",
    "requests>=2.32.3",
    "httpx>=0.27.2",
    "anyio>=4.9.0",
    "rich>=14.0.0",
    "click>=8.1.8",
]

[project.optional-dependencies]
memory = [
    "sentence-transformers>=4.0.2",
]
cli = [
    "rich>=14.0.0",
    "click>=8.1.8",
    "textual>=3.2.0",
]
all = [
    "browser-use-script[memory,cli]",
]

[project.urls]
Repository = "https://github.com/PierrunoYT/browser-use-script"

[project.scripts]
browser-use-cli = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "*.py",
    "README.md",
    "LICENSE",
    ".env.example",
]

[tool.ruff]
line-length = 130
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
